Reduce the size by removing top-level lib crate type
