+++ clis/nordvpnlite/Cargo.toml
@@ -21,7 +21,7 @@ interprocess = { version = "2.4.2", features = ["tokio"] }
 
 nix = { version = "0.31.3", features = ["signal"] }
 
-telio = { path = "../..", features = ["disable_ens"], default-features = false }
+telio-core = { workspace = true, features = ["disable_ens"], default-features = false }
 tokio.workspace = true
 signal-hook-tokio = { version = "0.4.0", features = ["futures-v0_3"] }
 futures.workspace = true
