+++ clis/tcli/Cargo.toml
@@ -35,7 +35,7 @@ thiserror.workspace = true
 time.workspace = true
 tokio = { workspace = true, features = ["full"] }
 
-telio = { path = "../.." }
+telio-core = { workspace = true, features = ["enable_ens"] }
 telio-crypto.workspace = true
 telio-model.workspace = true
 telio-proto.workspace = true
