+++ crates/telio-core/src/adapter.rs
