+++ crates/telio-core/src/lib.rs
@@ -1,0 +1,95 @@
+//! Core telio implementation: device orchestration and supporting types.
+
+pub mod adapter;
+pub mod defaults_builder;
+pub mod device;
+pub mod logging;
+pub mod types;
+
+pub use crate::types::*;
+
+/// Legacy alias for [`types`] — historically the top-level `telio` crate
+/// re-exported this module as `ffi_types`.
+pub mod ffi_types {
+    pub use crate::types::*;
+}
+
+/// cbindgen:ignore
+pub use telio_crypto as crypto;
+
+/// cbindgen:ignore
+pub use telio_dns;
+
+/// cbindgen:ignore
+pub use telio_proto;
+
+/// cbindgen:ignore
+pub use telio_proxy;
+
+/// cbindgen:ignore
+pub use telio_nurse;
+
+/// cbindgen:ignore
+pub use telio_relay;
+
+/// cbindgen:ignore
+pub use telio_traversal;
+
+/// cbindgen:ignore
+pub use telio_sockets;
+
+/// cbindgen:ignore
+pub use telio_task;
+
+/// cbindgen:ignore
+pub use telio_wg;
+
+/// cbindgen:ignore
+pub use telio_model;
+
+/// cbindgen:ignore
+pub use telio_utils;
+
+/// cbindgen:ignore
+pub use telio_lana;
+
+#[cfg(feature = "enable_firewall")]
+/// cbindgen:ignore
+pub use telio_firewall;
+
+use crate::logging::{HIDE_THREAD_ID_IN_LOGS, LOGGER_STOPPER, TIMESTAMPS_IN_LOGS};
+use telio_utils::telio_log_warn;
+use telio_wg::AdapterType;
+
+/// Set the global logger.
+/// # Parameters
+/// - `log_level`: Max log level to log.
+/// - `logger`: Callback to handle logging events.
+pub fn set_global_logger(log_level: TelioLogLevel, logger: Box<dyn TelioLoggerCb>) {
+    let tracing_subscriber = logging::build_subscriber(log_level, logger);
+    if tracing::subscriber::set_global_default(tracing_subscriber).is_err() {
+        telio_log_warn!("Could not set logger, because logger had already been set by previous libtelio instance");
+    }
+}
+
+/// Unset the global logger.
+/// After this call finishes, previously registered logger will not be called.
+pub fn unset_global_logger() {
+    LOGGER_STOPPER.stop();
+}
+
+/// Enable timestamps in logs generated by libtelio.
+/// When enabled host application doesn't need to add timestamps in the logging callback.
+pub fn add_timestamps_to_logs() {
+    TIMESTAMPS_IN_LOGS.store(true, std::sync::atomic::Ordering::Relaxed);
+}
+
+/// Control if thread IDs should be shown in the logs
+pub fn hide_thread_id_in_logs(hide: bool) {
+    HIDE_THREAD_ID_IN_LOGS.store(hide, std::sync::atomic::Ordering::Relaxed);
+}
+
+/// Get default recommended adapter type for platform.
+pub fn get_default_adapter() -> TelioAdapterType {
+    AdapterType::default().into()
+}
