+++ src/ffi.rs
@@ -1,7 +1,4 @@
-pub mod adapter;
-pub mod defaults_builder;
-pub mod logging;
-pub mod types;
+pub use telio_core::{adapter, defaults_builder, logging, types};
 
 use anyhow::anyhow;
 use ffi_helpers::{error_handling, panic as panic_handling};
@@ -21,13 +18,11 @@ use std::{
     time::Duration,
 };
 
-use self::{
-    logging::HIDE_THREAD_ID_IN_LOGS, logging::LOGGER_STOPPER, logging::TIMESTAMPS_IN_LOGS, types::*,
-};
-pub use crate::adapter::{
+use self::types::*;
+pub use telio_core::adapter::{
     CustomAdapter, TelioCustomAdapter, WgCmd, WgDevice, WgInterface, WgPeer, WgResponse,
 };
-use crate::device::{Device, DeviceConfig, Result as DevResult};
+use telio_core::device::{Device, DeviceConfig, Result as DevResult};
 use telio_model::{
     config::{Config, ConfigParseError},
     event::*,
@@ -39,7 +34,7 @@ use telio_model::{
 // debug tools
 use telio_utils::{
     backtrace::log_current_backtrace, commit_sha, telio_log_debug, telio_log_error, telio_log_info,
-    telio_log_warn, version_tag,
+    version_tag,
 };
 
 const DEFAULT_PANIC_MSG: &str = "libtelio panicked";
@@ -68,38 +63,6 @@ where
             Err(TelioError::UnknownError { inner: err_string })
         }
     }
-}
-
-/// Set the global logger.
-/// # Parameters
-/// - `log_level`: Max log level to log.
-/// - `logger`: Callback to handle logging events.
-pub fn set_global_logger(log_level: TelioLogLevel, logger: Box<dyn TelioLoggerCb>) {
-    let tracing_subscriber = logging::build_subscriber(log_level, logger);
-    if tracing::subscriber::set_global_default(tracing_subscriber).is_err() {
-        telio_log_warn!("Could not set logger, because logger had already been set by previous libtelio instance");
-    }
-}
-/// Unset the global logger.
-/// After this call finishes, previously registered logger will not be called.
-pub fn unset_global_logger() {
-    LOGGER_STOPPER.stop();
-}
-
-/// Enable timestamps in logs generated by libtelio.
-/// When enabled host application doesn't need to add timestamps in the logging callback.
-pub fn add_timestamps_to_logs() {
-    TIMESTAMPS_IN_LOGS.store(true, std::sync::atomic::Ordering::Relaxed);
-}
-
-/// Control if thread IDs should be shown in the logs
-pub fn hide_thread_id_in_logs(hide: bool) {
-    HIDE_THREAD_ID_IN_LOGS.store(hide, std::sync::atomic::Ordering::Relaxed);
-}
-
-/// Get default recommended adapter type for platform.
-pub fn get_default_adapter() -> TelioAdapterType {
-    AdapterType::default().into()
 }
 
 /// Get current version tag.
